left a little

+++ booksShop8/booksShop8/ViewModels/Books.cs
@@ -19,8 +19,8 @@ namespace booksShop8.ViewModels
         public string author { get; set; }
         public string genre { get; set; }
         public string status { get; set; }
-        public Nullable<int> bookMark { get; set; }
-        public Nullable<int> bookMarkCount { get; set; }
+        public Nullable<double> bookMark { get; set; }
+        public Nullable<double> bookMarkCount { get; set; }
         public Nullable<int> Quantities { get; set; }
         public byte[] bookImg { get; set; }
         public override bool Equals(object obj)
